+++ src/guard.rs
@@ -1,19 +1,20 @@
-use std::collections::VecDeque;
+use std::collections::{HashMap, VecDeque};
 use std::io;
 use std::net::SocketAddr;
 use std::pin::Pin;
 use std::task::{ready, Context, Poll};
+use std::time::Instant;
 
 use futures::Sink;
 use log::trace;
 use pin_project_lite::pin_project;
 
+use crate::estimator::{Estimator, RFC6298Impl};
 use crate::link::SharedLink;
 use crate::opts::FlushStrategy;
-use crate::packet::connected::{self, Frame, FrameSet, FramesRef};
+use crate::packet::connected::{self, AckOrNack, Frame, FrameSet, Frames, FramesRef, Record};
 use crate::packet::{Packet, FRAME_SET_HEADER_SIZE};
-use crate::resend_map::ResendMap;
-use crate::utils::u24;
+use crate::utils::{u24, ConnId, Reactor};
 use crate::{Peer, Role};
 
 pin_project! {
@@ -62,7 +63,7 @@ where
             peer,
             role,
             cap,
-            resend: ResendMap::new(role, peer),
+            resend: ResendMap::new(role, peer, Box::new(RFC6298Impl::new())),
         }
     }
 }
@@ -77,7 +78,7 @@ where
 
         this.link
             .process_ack()
-            .for_each(|ack| this.resend.on_ack(ack));
+            .for_each(|(ack, received_at)| this.resend.on_ack(ack, received_at));
         this.link
             .process_nack()
             .for_each(|nack| this.resend.on_nack_into(nack, this.buf));
@@ -269,4 +270,285 @@ where
     }
 }
 
-// TODO: test
+struct ResendEntry {
+    frames: Option<Frames>,
+    send_at: Instant,
+    expired_at: Instant,
+}
+
+struct ResendMap {
+    map: HashMap<u24, ResendEntry>,
+    role: Role,
+    peer: Peer,
+    last_record_expired_at: Instant,
+    estimator: Box<dyn Estimator + Send>,
+}
+
+impl ResendMap {
+    fn new(role: Role, peer: Peer, estimator: Box<dyn Estimator + Send>) -> Self {
+        Self {
+            map: HashMap::new(),
+            role,
+            peer,
+            last_record_expired_at: Instant::now(),
+            estimator,
+        }
+    }
+
+    fn record(&mut self, seq_num: u24, frames: Frames) {
+        let now = Instant::now();
+        self.map.insert(
+            seq_num,
+            ResendEntry {
+                frames: Some(frames),
+                send_at: now,
+                expired_at: now + self.estimator.rto(),
+            },
+        );
+    }
+
+    fn on_ack(&mut self, ack: AckOrNack, received_at: Instant) {
+        for record in ack.records {
+            match record {
+                Record::Range(start, end) => {
+                    for i in start.to_u32()..=end.to_u32() {
+                        if let Some(ResendEntry { send_at, .. }) = self.map.remove(&i.into()) {
+                            let rtt = received_at.saturating_duration_since(send_at);
+                            self.estimator.update(rtt);
+                            trace!(
+                                "[{}] seq_num {i} is ACKed by {}, RTT: {rtt:?}, estimated RTO: {:?}",
+                                self.role,
+                                self.peer,
+                                self.estimator.rto()
+                            );
+                        }
+                    }
+                }
+                Record::Single(seq_num) => {
+                    if let Some(ResendEntry { send_at, .. }) = self.map.remove(&seq_num) {
+                        let rtt = received_at.saturating_duration_since(send_at);
+                        self.estimator.update(rtt);
+                        trace!(
+                            "[{}] seq_num {seq_num} is ACKed by {}, RTT: {rtt:?}, estimated RTO: {:?}",
+                            self.role,
+                            self.peer,
+                            self.estimator.rto()
+                        );
+                    }
+                }
+            }
+        }
+    }
+
+    fn on_nack_into(&mut self, nack: AckOrNack, buffer: &mut VecDeque<Frame>) {
+        trace!("[{}] receive NACKs {nack:?} from {}", self.role, self.peer);
+        for record in nack.records {
+            match record {
+                Record::Range(start, end) => {
+                    for i in start.to_u32()..=end.to_u32() {
+                        if let Some(entry) = self.map.remove(&i.into()) {
+                            buffer.extend(entry.frames.unwrap());
+                        }
+                    }
+                }
+                Record::Single(seq_num) => {
+                    if let Some(entry) = self.map.remove(&seq_num) {
+                        buffer.extend(entry.frames.unwrap());
+                    }
+                }
+            }
+        }
+    }
+
+    /// `process_stales` collect all stale frames into buffer and remove the expired entries
+    fn process_stales(&mut self, buffer: &mut VecDeque<Frame>) {
+        let now = Instant::now();
+        if now < self.last_record_expired_at {
+            trace!(
+                "[{}] skip scanning the resend map, last record expired at {:?}",
+                self.role,
+                self.last_record_expired_at - now
+            );
+            return;
+        }
+        // find the first expired_at larger than now
+        let mut min_expired_at = now + self.estimator.rto();
+        let len_before = self.map.len();
+        self.map.retain(|_, entry| {
+            if entry.expired_at <= now {
+                buffer.extend(entry.frames.take().unwrap());
+                false
+            } else {
+                min_expired_at = min_expired_at.min(entry.expired_at);
+                true
+            }
+        });
+        debug_assert!(min_expired_at > now);
+        self.last_record_expired_at = min_expired_at;
+
+        let len = self.map.len();
+        if len_before > len {
+            // clear the estimator if detected packet loss
+            self.estimator.clear();
+        }
+        trace!(
+            "[{}]: resend {} stales, {} entries remains",
+            self.role,
+            len_before - len,
+            len,
+        );
+    }
+
+    fn is_empty(&self) -> bool {
+        self.map.is_empty()
+    }
+
+    /// `poll_wait` suspends the task when the resend map needs to wait for the next resend
+    fn poll_wait(&self, cx: &mut Context<'_>) -> Poll<()> {
+        let expired_at;
+        let seq_num;
+        let now = Instant::now();
+        if let Some((seq, entry)) = self.map.iter().min_by_key(|(_, entry)| entry.expired_at)
+            && entry.expired_at > now
+        {
+            expired_at = entry.expired_at;
+            seq_num = *seq;
+        } else {
+            return Poll::Ready(());
+        }
+        let c_id = ConnId::new(self.role.guid(), self.peer.guid);
+        trace!(
+            "[{}]: wait on {c_id:?} for resend seq_num {} to {} within {:?}",
+            self.role,
+            seq_num,
+            self.peer,
+            expired_at - now
+        );
+        Reactor::get().insert_timer(c_id, expired_at, cx.waker());
+        Poll::Pending
+    }
+}
+
+#[cfg(test)]
+mod test {
+    use std::collections::VecDeque;
+    use std::task::{Context, Poll};
+    use std::time::{Duration, Instant};
+
+    use bytes::Bytes;
+
+    use super::ResendMap;
+    use crate::estimator::RFC6298Impl;
+    use crate::packet::connected::{AckOrNack, Flags, Frame};
+    use crate::utils::tests::{test_trace_log_setup, TestWaker};
+    use crate::{Peer, Reliability, Role};
+
+    const TEST_RTO: Duration = Duration::from_millis(1200);
+
+    #[test]
+    fn test_resend_map_works() {
+        let mut map = ResendMap::new(
+            Role::test_server(),
+            Peer::test(),
+            Box::new(RFC6298Impl::new()),
+        );
+        map.record(0.into(), vec![]);
+        map.record(1.into(), vec![]);
+        map.record(2.into(), vec![]);
+        map.record(3.into(), vec![]);
+        assert!(!map.is_empty());
+        map.on_ack(
+            AckOrNack::extend_from([0, 1, 2, 3].into_iter().map(Into::into), 100).unwrap(),
+            Instant::now(),
+        );
+        assert!(map.is_empty());
+
+        map.record(
+            4.into(),
+            vec![Frame {
+                flags: Flags::new(Reliability::Unreliable, false),
+                reliable_frame_index: None,
+                seq_frame_index: None,
+                ordered: None,
+                fragment: None,
+                body: Bytes::from_static(b"1"),
+            }],
+        );
+        map.record(
+            5.into(),
+            vec![
+                Frame {
+                    flags: Flags::new(Reliability::Unreliable, false),
+                    reliable_frame_index: None,
+                    seq_frame_index: None,
+                    ordered: None,
+                    fragment: None,
+                    body: Bytes::from_static(b"2"),
+                },
+                Frame {
+                    flags: Flags::new(Reliability::Unreliable, false),
+                    reliable_frame_index: None,
+                    seq_frame_index: None,
+                    ordered: None,
+                    fragment: None,
+                    body: Bytes::from_static(b"3"),
+                },
+            ],
+        );
+        let mut buffer = VecDeque::default();
+        map.on_nack_into(
+            AckOrNack::extend_from([4, 5].into_iter().map(Into::into), 100).unwrap(),
+            &mut buffer,
+        );
+        assert!(map.is_empty());
+        assert_eq!(buffer.len(), 3);
+        assert_eq!(buffer.pop_front().unwrap().body, Bytes::from_static(b"1"));
+        assert_eq!(buffer.pop_front().unwrap().body, Bytes::from_static(b"2"));
+        assert_eq!(buffer.pop_front().unwrap().body, Bytes::from_static(b"3"));
+    }
+
+    #[test]
+    fn test_resend_map_stales() {
+        let mut map = ResendMap::new(
+            Role::test_server(),
+            Peer::test(),
+            Box::new(RFC6298Impl::new()),
+        );
+        map.record(0.into(), vec![]);
+        map.record(1.into(), vec![]);
+        map.record(2.into(), vec![]);
+        std::thread::sleep(TEST_RTO);
+        map.record(3.into(), vec![]);
+        let mut buffer = VecDeque::default();
+        map.process_stales(&mut buffer);
+        assert_eq!(map.map.len(), 1);
+    }
+
+    #[tokio::test]
+    async fn test_resend_map_poll_wait() {
+        let _guard = test_trace_log_setup();
+
+        let mut map = ResendMap::new(
+            Role::test_server(),
+            Peer::test(),
+            Box::new(RFC6298Impl::new()),
+        );
+        map.record(0.into(), vec![]);
+        std::thread::sleep(TEST_RTO);
+        map.record(1.into(), vec![]);
+        map.record(2.into(), vec![]);
+        map.record(3.into(), vec![]);
+
+        let mut buffer = VecDeque::default();
+
+        let res = map.poll_wait(&mut Context::from_waker(&TestWaker::create()));
+        assert!(matches!(res, Poll::Ready(_)));
+
+        map.process_stales(&mut buffer);
+        assert_eq!(map.map.len(), 3);
+
+        std::future::poll_fn(|cx| map.poll_wait(cx)).await;
+        map.process_stales(&mut buffer);
+        assert!(map.map.len() < 3);
+    }
+}